+++ bot-app/config/config.py
@@ -5,7 +5,12 @@ from urllib.parse import quote
 
 from pydantic import BaseModel, PostgresDsn, ValidationError, computed_field
 from pydantic_core import MultiHostUrl
-from pydantic_settings import BaseSettings, SettingsConfigDict
+from pydantic_settings import (
+    BaseSettings,
+    PydanticBaseSettingsSource,
+    SettingsConfigDict,
+    YamlConfigSettingsSource,
+)
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.INFO)
@@ -14,6 +19,7 @@ PROJECT_DIR = Path(__file__).resolve().parent.parent
 
 class BotConfig(BaseModel):
     token: str
+    owner_tg_id: int
 
 
 class DatabaseConfig(BaseModel):
@@ -53,6 +59,7 @@ class RabbitMQConfig(BaseModel):
     username: str
     password: str
     vhostname: str
+    secure: bool = True
 
     @computed_field  # type: ignore[prop-decorator]
     @property
@@ -61,19 +68,58 @@ class RabbitMQConfig(BaseModel):
         safe_password = quote(self.password, safe="")
         safe_vhost = quote(self.vhostname, safe="")
         domain = quote(self.host.encode("idna").decode())
+        protocol = "amqps" if self.secure else "amqp"
 
-        return (
-            f"amqps://{safe_username}:{safe_password}@{domain}:{self.port}/{safe_vhost}"
-        )
+        return f"{protocol}://{safe_username}:{safe_password}@{domain}:{self.port}/{safe_vhost}"
 
 
 class Settings(BaseSettings):
     model_config = SettingsConfigDict(
-        env_file=(PROJECT_DIR / ".env.template", PROJECT_DIR / ".env"),
+        env_file=(
+            PROJECT_DIR / ".env.template",
+            PROJECT_DIR / ".env",
+        ),
         case_sensitive=False,
         env_nested_delimiter="__",
         env_prefix="APP_CONFIG__",
+        yaml_file=(
+            PROJECT_DIR / "config.default.yaml",
+            PROJECT_DIR / "config.local.yaml",
+        ),
+        yaml_config_section="tasks-bot",
     )
+
+    @classmethod
+    def settings_customise_sources(
+        cls,
+        settings_cls: type[BaseSettings],
+        init_settings: PydanticBaseSettingsSource,
+        env_settings: PydanticBaseSettingsSource,
+        dotenv_settings: PydanticBaseSettingsSource,
+        file_secret_settings: PydanticBaseSettingsSource,
+    ) -> tuple[PydanticBaseSettingsSource, ...]:
+        """
+        Define the sources and their order for loading the settings values.
+
+        Args:
+            settings_cls: The Settings class.
+            init_settings: The `InitSettingsSource` instance.
+            env_settings: The `EnvSettingsSource` instance.
+            dotenv_settings: The `DotEnvSettingsSource` instance.
+            file_secret_settings: The `SecretsSettingsSource` instance.
+
+        Returns:
+            A tuple containing the sources and their order
+            for loading the settings values.
+        """
+        return (
+            init_settings,
+            env_settings,
+            dotenv_settings,
+            file_secret_settings,
+            YamlConfigSettingsSource(settings_cls),
+        )
+
     bot: BotConfig
     db: DatabaseConfig
     rabbit: RabbitMQConfig